requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-session==0.5.0
pg8000==1.30.4
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
